Ask user for permission before installing required packages
